requirements.txt
fastapi
uvicorn
google-api-python-client
python-dotenv
google-generativeai
langchain-google-genai
chromadb
langchain
pandas
langchain-community
jq
youtube-transcript-api
bs4
youtube-search
praw
